fix/restore js libs, which got accidently reverted to an earlier state

+++ assets/js/netresearch/model/Contract.js
@@ -5,13 +5,13 @@ Ext.define('Netresearch.model.Contract', {
         {name: 'user_id', type: 'integer'},
         {name: 'start', type: 'date', dateFormat: 'Y-m-d'},
         {name: 'end', type: 'date', dateFormat: 'Y-m-d'},
-        {name: 'hours_0', type: 'integer'},
-        {name: 'hours_1', type: 'integer'},
-        {name: 'hours_2', type: 'integer'},
-        {name: 'hours_3', type: 'integer'},
-        {name: 'hours_4', type: 'integer'},
-        {name: 'hours_5', type: 'integer'},
-        {name: 'hours_6', type: 'integer'},
-        {name: 'hours_7', type: 'integer'}
+        {name: 'hours_0', type: 'float'},
+        {name: 'hours_1', type: 'float'},
+        {name: 'hours_2', type: 'float'},
+        {name: 'hours_3', type: 'float'},
+        {name: 'hours_4', type: 'float'},
+        {name: 'hours_5', type: 'float'},
+        {name: 'hours_6', type: 'float'},
+        {name: 'hours_7', type: 'float'}
     ]
 });
